# server/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Status": {
				"enum": [
					"ordering",
					"ordered",
					"delivery",
					"canceled",
					"completed"
				],
				"type": "string"
			},
			"PaymentMethod": {
				"enum": [
					"cod",
					"banktransfer",
					"e-wallet",
					"gateway"
				],
				"type": "string"
			},
...
			"Change": {
				"enum": [
					"DESC",
					"ASC"
				],
				"type": "string"
			},
			"UserRole": {
				"enum": [
					"all",
					"admin",
					"editor",
					"customer"
				],
				"type": "string"
			},
...
			"IOrderDetail": {
				"properties": {
					"orderId": {
						"type": "number",
						"format": "double"
					},
					"productVariantId": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unitPrice": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"productVariantId",
					"quantity",
					"unitPrice",
					"id",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"BlogCategory": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Blog": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"title": {
						"type": "string"
					},
					"categoryId": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"$ref": "#/components/schemas/BlogCategory"
					},
					"imgPath": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"commentAllow": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"createdAt",
					"title",
					"categoryId",
					"category",
					"imgPath",
					"content",
					"commentAllow"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBlogs": {
				"properties": {
					"blogs": {
						"items": {
							"$ref": "#/components/schemas/Blog"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"blogs",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BlogField": {
				"enum": [
					"title",
					"createdAt"
				],
				"type": "string"
			},
			"ValidateBlogModel": {
				"properties": {
					"title": {
						"type": "string",
						"pattern": "^(?!\\s*$).+"
					},
					"categoryId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"imgPath": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"commentAllow": {
						"type": "boolean"
					}
				},
				"required": [
					"title",
					"categoryId",
					"content",
					"commentAllow"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBlogCommentCreateProps": {
				"properties": {
					"text": {
						"type": "string"
					},
					"blogId": {
						"type": "number",
						"format": "double"
					},
					"parentCommentId": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"text",
					"blogId",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IOrder": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"paymentMethod": {
						"$ref": "#/components/schemas/PaymentMethod"
					},
					"paymentCheck": {
						"type": "boolean"
					},
					"address": {
						"type": "string"
					},
					"shipping": {
						"type": "number",
						"format": "double"
					},
					"orderAt": {
						"type": "string",
						"format": "date-time"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/IOrderDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"userId",
					"id",
					"createdAt",
					"status",
					"total",
					"paymentMethod",
					"paymentCheck",
					"address",
					"shipping",
					"orderAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IUser": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"orders": {
						"items": {
							"$ref": "#/components/schemas/IOrder"
						},
						"type": "array"
					},
					"fname": {
						"type": "string"
					},
					"lname": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"fname",
					"lname",
					"id",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBlogComment": {
				"properties": {
					"text": {
						"type": "string"
					},
					"blogId": {
						"type": "number",
						"format": "double"
					},
					"parentCommentId": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"childComments": {
						"items": {
							"$ref": "#/components/schemas/IBlogCommentCreateProps"
						},
						"type": "array"
					},
					"user": {
						"$ref": "#/components/schemas/IUser"
					}
				},
				"required": [
					"text",
					"blogId",
					"userId",
					"id",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBlogCommentUpdateProps": {
				"properties": {
					"text": {
						"type": "string"
					},
					"blogId": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"text",
					"blogId",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ValidateBlogCateModel": {
				"properties": {
					"name": {
						"type": "string",
						"pattern": "^(?!\\s*$).+"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Brand": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"imgPath": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"name",
					"imgPath"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateBrandModel": {
				"properties": {
					"name": {
						"type": "string",
						"pattern": "^(?!\\s*$).+"
					},
					"imgPath": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"in": "cookie",
				"name": "token"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "belle-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "LinhTran"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/blogs": {
			"get": {
				"operationId": "GetBlogs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBlogs"
								}
							}
						}
					}
				},
				"description": "Get all blogs. Search blogs by its category, keyword. Get specific number of blogs by time in descending order",
				"tags": [
					"Blog"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BlogField"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "change",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Change"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateBlog",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Blog"
								}
							}
						}
					}
				},
				"description": "Create a new blog",
				"tags": [
					"Blog"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateBlogModel"
							}
						}
					}
				}
			}
		},
		"/blogs/{id}": {
			"get": {
				"operationId": "GetBlogById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Blog"
								}
							}
						}
					}
				},
				"description": "Get details for a blog by its id",
				"tags": [
					"Blog"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateBlogById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Blog"
								}
							}
						}
					}
				},
				"description": "Update a blog by its id",
				"tags": [
					"Blog"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateBlogModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBlogById",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete a blog  by its id",
				"tags": [
					"Blog"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			}
		},
		"/blogs/{blogId}/comments": {
			"get": {
				"operationId": "GetCommentsOfBlog",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IBlogComment"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get blog comments by blogId, default number of comments is 2",
				"tags": [
					"Blog"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "blogId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"in": "query",
						"name": "date",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					}
				]
			}
		},
		"/blogs/comments": {
			"post": {
				"operationId": "CreateComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBlogComment"
								}
							}
						}
					}
				},
				"description": "Create new blog comment",
				"tags": [
					"Blog"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor",
							"customer"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IBlogCommentCreateProps"
							}
						}
					}
				}
			}
		},
		"/blogs/comments/{commentId}": {
			"patch": {
				"operationId": "UpdateCommentById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBlogComment"
								}
							}
						}
					}
				},
				"description": "Update a blog comment by its Id",
				"tags": [
					"Blog"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor",
							"customer"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "commentId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IBlogCommentUpdateProps"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCommentById",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete a blog comment by its Id",
				"tags": [
					"Blog"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor",
							"customer"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "commentId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			}
		},
...
		"/blog-categories": {
			"get": {
				"operationId": "GetBlogCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BlogCategory"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all blog categories",
				"tags": [
					"Blog Category"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateBlogCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlogCategory"
								}
							}
						}
					}
				},
				"description": "Create a new blog category",
				"tags": [
					"Blog Category"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateBlogCateModel"
							}
						}
					}
				}
			}
		},
		"/blog-categories/{id}": {
			"patch": {
				"operationId": "UpdateBlogCategoryById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlogCategory"
								}
							}
						}
					}
				},
				"description": "Update a blog category by its id",
				"tags": [
					"Blog Category"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateBlogCateModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBlogCategoryById",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete a blog category by its id",
				"tags": [
					"Blog Category"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			}
		},
...
		"/brands": {
			"get": {
				"operationId": "GetBrands",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Brand"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all brands",
				"tags": [
					"Product Brand"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateBrand",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Brand"
								}
							}
						}
					}
				},
				"description": "Create new brand",
				"tags": [
					"Product Brand"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateBrandModel"
							}
						}
					}
				}
			}
		},
		"/brands/{id}": {
			"patch": {
				"operationId": "UpdateBrandById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Brand"
								}
							}
						}
					}
				},
				"description": "Update brand info",
				"tags": [
					"Product Brand"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateBrandModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBrandById",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete brand",
				"tags": [
					"Product Brand"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}